Fix latent bugs, dedupe logic, and split oversized test modules
Bug fixes:
- tile: `is_sequential_to` underflowed (u32) when the argument was 1m,
  panicking in debug builds; rewritten with same-suit + abs_diff
- hand: kan melds displayed `tiles[0]` as the 4th tile, duplicating the
  red 5 in concealed kans; now uses the new `Meld::kan_fourth_tile`
- cpu: shanten calculations in `evaluate_discards`, `count_acceptance`,
  `should_attack`, `call_reduces_shanten_pon/chi` and the legacy ankan
  check ignored melds, so open hands were judged several tiles further
  from tenpai than they were (e.g. an open tenpai hand always folded in
  the endgame, and pon/chi comparisons were asymmetric)

Dedup / readability:
- core: add `Meld::kan_fourth_tile` / `Meld::expanded_tiles` and use them
  in hand display, round events, and dora counting (removes two copies of
  `kan_fourth_tile`)
- round: extract `announce_discard_and_check_calls` (shared by
  `do_discard` / `do_riichi`) and `push_draw_events` (shared by
  `do_draw` / `draw_after_kan`)
- wall: draw from a VecDeque instead of `Vec::remove(0)`
- cpu: simplify `is_yakuhai` / `get_yakuhai_types` using Wind discriminants,
  drop the empty daiminkan branch
- fu: reuse `Tile::is_1_9_honor`; table: reuse `total_rounds()`
- checker: fix a misplaced junchan comment

Module size:
- move unit tests out of round/mod.rs (3077 -> 1789 lines),
  cpu/heuristics.rs (3615 -> 1572) and cpu/client.rs (2303 -> 776)
  into dedicated test files

Adds regression tests for the underflow and the meld-aware shanten fixes.
Verified: cargo test (604 passed), clippy/fmt clean, wasm check, and the
100-game simulation_regression run (strength ordering preserved).
